Improved context script
